Add debug-intent LC-transition test
Add smoke_test_fc_debug_intent_lc_transition: under the strap a
zero-token TEST_UNLOCKED0->TEST_LOCKED0 succeeds while a token-
dependent transition is blocked (the SECRET_LC_TRANSITION token is
suppressed). Drop DEV->SCRAP from the zeroize test (SCRAP asserts
lc_escalate_en) and keep the LC state readback. Remove the redundant
program-blocked test, now covered by the unlocked-FE3 write check.
