+++ src/integration/stimulus/testsuites/caliptra_ss_master_test_list.csv
@@ -141,4 +141,4 @@ $CALIPTRA_SS_ROOT/src/integration/test_suites/smoke_test_fc_filter_rule_write_ze
 $CALIPTRA_SS_ROOT/src/integration/test_suites/caliptra_ss_lcc_volatile_unlock_with_state_error/caliptra_ss_lcc_volatile_unlock_with_state_error , Directed       , Nightly       , None, L1  , caliptra_ss_top_tb, None           , None     , 100
 $CALIPTRA_SS_ROOT/src/integration/test_suites/smoke_test_fc_debug_intent_secret_zeroize/smoke_test_fc_debug_intent_secret_zeroize                 , Directed       , Nightly       , None, L1  , caliptra_ss_top_tb, None           , None     , 10
 $CALIPTRA_SS_ROOT/src/integration/test_suites/smoke_test_fc_debug_intent_secret_zeroize_baseline/smoke_test_fc_debug_intent_secret_zeroize_baseline, Directed       , Nightly       , None, L1  , caliptra_ss_top_tb, None           , None     , 10
-$CALIPTRA_SS_ROOT/src/integration/test_suites/smoke_test_fc_debug_intent_secret_program_blocked/smoke_test_fc_debug_intent_secret_program_blocked, Directed       , Nightly       , None, L1  , caliptra_ss_top_tb, None           , None     , 10
+$CALIPTRA_SS_ROOT/src/integration/test_suites/smoke_test_fc_debug_intent_lc_transition/smoke_test_fc_debug_intent_lc_transition                  , Directed       , Nightly       , None, L1  , caliptra_ss_top_tb, None           , None     , 10
