+++ src/integration/test_suites/smoke_test_fc_debug_intent_lc_transition/smoke_test_fc_debug_intent_lc_transition.yml
@@ -14,17 +14,18 @@
 #
 ---
 seed: 1
-testname: smoke_test_fc_debug_intent_secret_program_blocked
+testname: smoke_test_fc_debug_intent_lc_transition
 pre-exec: |
-  echo "Running pre_exec for [smoke_test_fc_debug_intent_secret_program_blocked]"
+  echo "Running pre_exec for [smoke_test_fc_debug_intent_lc_transition]"
   rm -rf otp-img.2048.vmem &&
   $CALIPTRA_SS_ROOT/tools/scripts/fuse_ctrl_script/gen_fuse_ctrl_vmem.py \
-      --lc-state 16 \
+      --lc-state 1 \
+      --lc-cnt 0 \
       --lc-state-def $CALIPTRA_SS_ROOT/tools/scripts/fuse_ctrl_script/lc_ctrl_state.hjson \
       --mmap-def $CALIPTRA_SS_ROOT/src/fuse_ctrl/data/otp_ctrl_mmap.hjson \
-      --token-configuration $CALIPTRA_SS_ROOT/src/integration/test_suites/smoke_test_fc_debug_intent_secret_program_blocked/test_unlock_token.hjson \
+      --token-configuration $CALIPTRA_SS_ROOT/src/integration/test_suites/smoke_test_fc_debug_intent_lc_transition/test_unlock_token.hjson \
       -o otp-img.2048.vmem &&
   CALIPTRA_ROOT=$CALIPTRA_SS_ROOT/third_party/caliptra-rtl make -f $CALIPTRA_SS_ROOT/third_party/caliptra-rtl/tools/scripts/Makefile CALIPTRA_INTERNAL_TRNG=0 TESTNAME=smoke_test_mbox program.hex &&
-  make -f $CALIPTRA_SS_ROOT/tools/scripts/Makefile TESTNAME=smoke_test_fc_debug_intent_secret_program_blocked mcu_program.hex
+  make -f $CALIPTRA_SS_ROOT/tools/scripts/Makefile TESTNAME=smoke_test_fc_debug_intent_lc_transition mcu_program.hex
 plusargs:
     - '+CALIPTRA_SS_DEBUG_INTENT'
